change PriorityCollection to use a ConcurrentSkipListSet internally
There are no alternatives to PriorityCollection for standard use of the
API, so it must be threadsafe.

+++ api/src/main/java/org/semanticweb/owlapi/annotations/HasPriority.java
@@ -30,5 +30,5 @@ import java.lang.annotation.RetentionPolicy;
 public @interface HasPriority {
 
     /** @return priority value for this object */
-    double value();
+    int value();
 }
